Added "@container()" alias in the template
The template has @containers. These are regions that output a view.
Each view can be assigned to Containers. It's much like how Joomla
module or Drupal blocks work. Instead here, it's called container.

Syntax in the Theme: <?=@container('containername', 'chrome')?>

The chrome is the wrapper where the view is rendered.

On a side note, we also made the theme view a singleton.